Vive Mr saradayan qui m'a créé des entity pour pouvoir sauvegarder des coo

+++ src/main/java/com/sp/service/FireSimulationService.java
@@ -91,7 +91,9 @@ public class FireSimulationService {
 		FireDto fireDtoToRet = null;
 		FireDto[] listFire = this.getFire();
 		for (FireDto fireDto: listFire) {
-			if (fireDto.getId()==_fireId) {
+			//System.out.println("DTO "+fireDto.getId());
+			//System.out.println("PAram"+_fireId);
+			if (fireDto.getId().intValue()==_fireId) {
 				fireDtoToRet=fireDto;
 				break;
 			}
